+++ src/Root/App/styles.css
@@ -53,6 +53,7 @@
 
     /* COLOR TEXT */
     --color-text: rgba(0, 0, 0, .7);
+    --color-text-label: rgba(0, 0, 0, .5);
     --color-text-on-primary: rgba(255, 255, 255, .9);
     --color-text-on-accent: rgba(255, 255, 255, .9);
     --color-text-on-danger: rgba(255, 255, 255, .9);
