+++ src/components/EntryForm/AnalysisInput.tsx
