+++ src/components/EntryForm/DetailsInput.tsx
