+++ src/components/EntryForm/FigureInput.tsx
