+++ src/components/EntryForm/StrataInput.tsx
