+++ src/components/EntryForm/styles.css
@@ -1,0 +1,142 @@
+.empty-message {
+    padding: var(--spacing-medium);
+    color: rgba(0, 0, 0, .3);
+}
+
+.block {
+    padding: var(--spacing-medium);
+}
+
+.strata-input,
+.age-input {
+    display: flex;
+    align-items: flex-end;
+    flex-wrap: wrap;
+    padding: 0 calc(var(--spacing-medium) - var(--spacing-small));
+
+    >* {
+        margin: 0 var(--spacing-small);
+    }
+
+    button {
+        margin-bottom: var(--spacing-medium);
+    }
+}
+
+.row {
+    display: flex;
+    padding: var(--spacing-small) var(--spacing-medium);
+
+    >* {
+        flex-grow: 1;
+    }
+}
+
+.process-url-button {
+    align-self: flex-end;
+    flex-grow: unset;
+    flex-shrink: 0;
+    margin-left: var(--spacing-medium);
+}
+
+.event-row {
+    display: flex;
+    align-items: flex-end;
+    padding: var(--spacing-small) calc(var(--spacing-medium) - var(--spacing-small));
+
+    .event-select-input {
+        flex-grow: 1;
+        margin: 0 var(--spacing-small);
+    }
+
+    .add-event-button {
+        flex-shrink: 0;
+        margin: var(--spacing-medium) var(--spacing-small);
+    }
+}
+
+.two-column-row {
+    display: flex;
+    padding: 0 calc(var(--spacing-medium) - var(--spacing-small));
+
+    >* {
+        flex-basis: calc(50% - 2 * var(--spacing-small));
+        margin: 0 var(--spacing-small);
+    }
+}
+
+.three-column-row {
+    display: flex;
+    padding: 0 calc(var(--spacing-medium) - var(--spacing-small));
+
+    >* {
+        flex-basis: calc(100% / 3 - 2 * var(--spacing-small));
+        margin: 0 var(--spacing-small);
+    }
+}
+
+.actions {
+    display: flex;
+    padding: var(--spacing-medium);
+
+    >* {
+        margin-left: var(--spacing-small);
+    }
+}
+
+.entry-form {
+    display: flex;
+    flex-direction: column;
+    background-color: var(--color-foreground);
+    overflow: auto;
+
+    .content {
+        display: flex;
+        flex-direction: column;
+        flex-grow: 1;
+        overflow: auto;
+
+        .tab-list {
+            display: flex;
+            flex-shrink: 0;
+
+            >* {
+                flex-basis: 0;
+                flex-grow: 1;
+                padding: var(--spacing-medium);
+            }
+        }
+
+        .details {
+            flex-grow: 1;
+            margin-bottom: var(--spacing-medium);
+            padding: var(--spacing-large);
+            overflow: auto;
+        }
+
+        .analysis-and-figures {
+            flex-grow: 1;
+            margin-bottom: var(--spacing-medium);
+            padding: var(--spacing-medium);
+            overflow: auto;
+        }
+
+        .review {
+            flex-grow: 1;
+            margin-bottom: var(--spacing-medium);
+            padding: var(--spacing-medium);
+            overflow: auto;
+
+            .review-assignee-input {
+                flex-grow: 1;
+            }
+        }
+    }
+}
+
+.add-event-modal {
+    .body {
+        padding-top: 0;
+        padding-bottom: 0;
+    }
+}
